refactor(editor): split schema resolution concerns
